[+ Orders specifying an unexisting customer will create it (email check).]

+++ src/main/java/com/acme/api/services/EmployeeService.java
@@ -39,7 +39,7 @@ public class EmployeeService implements EmployeeInterface{
 
     @Override
     public void deleteEmployee(long id) {
-        employeeRepository.deleteById((long) id);
+        employeeRepository.deleteById(id);
     }
 
     @Override
